reorganize directory structure for delayed_request module
