+++ src/cc_catalog_airflow/dags/provider_api_scripts/__init__.py
@@ -1,0 +1,7 @@
+import os
+import sys
+
+sys.path.insert(
+    0,
+    os.path.abspath(os.path.dirname(__file__))
+)
