+++ src/cc_catalog_airflow/dags/provider_api_scripts/common/storage/constants.py
