+++ src/cc_catalog_airflow/dags/provider_api_scripts/common/storage/test_columns.py
@@ -1,7 +1,7 @@
 import logging
 import string
 
-from storage import columns
+from common.storage import columns
 
 logging.basicConfig(
     format='%(asctime)s - %(name)s - %(levelname)s:  %(message)s',
