+++ src/cc_catalog_airflow/dags/provider_api_scripts/common/storage/test_constants.py
@@ -1,4 +1,4 @@
-from storage import constants
+from common.storage import constants
 
 
 def test_get_license_path_map_comprehenseion(monkeypatch):
