+++ src/cc_catalog_airflow/dags/provider_api_scripts/common/storage/test_util.py
@@ -1,6 +1,6 @@
 import logging
 
-from storage import util
+from common.storage import util
 
 logging.basicConfig(
     format='%(asctime)s - %(name)s - %(levelname)s:  %(message)s',
